+++ zk/close_resp_test.go
@@ -10,7 +10,7 @@ func TestDecodeCloseResp(t *testing.T) {
 	bytes := hex2Bytes(t, "00000003000000000000000700000000")
 	resp, err := DecodeCloseResp(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 3, resp.TransactionId)
+	assert.Equal(t, int32(3), resp.TransactionId)
 	assert.Equal(t, int64(7), resp.ZxId)
 	assert.Equal(t, EC_OK, resp.Error)
 }
