+++ zk/delete_req_test.go
@@ -10,7 +10,7 @@ func TestDecodeDeleteReq(t *testing.T) {
 	bytes := hex2Bytes(t, "00000001000000020000000c2f7a6b2d6e6f74666f756e6400000000")
 	req, err := DecodeDeleteReq(bytes)
 	assert.Nil(t, err)
-	assert.Equal(t, 1, req.TransactionId)
+	assert.Equal(t, int32(1), req.TransactionId)
 	assert.Equal(t, OP_DELETE, req.OpCode)
 	assert.Equal(t, "/zk-notfound", req.Path)
 	assert.Equal(t, 0, req.Version)
